+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-08T19:14:48.554Z",
+    "2026-03-09T05:19:42.278Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {