+++ packages/orq-rc/RELEASES.md
@@ -7519,3 +7519,13 @@ Based on:
 - [typescript v4.5.0-rc.39] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.39 - packages/orq-rc
+
+## 2026-03-09 05:22:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.749.1 (2.855.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.40] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.40 - packages/orq-rc